% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{get.cond.asym.surv.given.data}
\alias{get.cond.asym.surv.given.data}
\title{Asymptotic log-rank test (deprecated)}
\usage{
get.cond.asym.surv.given.data(surv.data)
}
\arguments{
\item{surv.data}{A data.frame containing three columns:
Survival - the event time.
Death - binary value. 1 is the event time is observed, 0 if censored.
cluster - variable indicating the group of each sample.}
}
\value{
asymptotic log-rank p-value.
}
\description{
Get the p-value for the asymptotic log-rank test. The test is performed using the survival
package survdiff function. This function is deprecated, and survdiff should be used for the asymptotic
log-rank test.
}
\examples{
get.cond.asym.surv.given.data(surv.data)
}
