% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{get.cond.perm.surv.given.data}
\alias{get.cond.perm.surv.given.data}
\title{Heinze log-rank test (deprecated)}
\usage{
get.cond.perm.surv.given.data(
  surv.data,
  max.num.perms = NULL,
  verbose = T,
  mc.cores = 50,
  seed = 42
)
}
\arguments{
\item{surv.data}{A data.frame containing three columns:
Survival - the event time.
Death - binary value. 1 is the event time is observed, 0 if censored.
cluster - variable indicating the group of each sample.}

\item{max.num.perms}{If this number is set, max.num.perms are performed.}

\item{verbose}{True <=> info on the number of permutations should be printed.}

\item{mc.cores}{Number of cores to use to parallelize permutations.}

\item{seed}{Initialization of R's random seed, or NULL to not set the seed.}
}
\value{
A list with the following entries:
pvalue - the estimation for log-rank's p-value.
conf.int - 95\% confidence interval for log-rank's p-value.
total.num.perms - number of permutations performed to estimate the p-value.
total.num.extreme.pvals - the number of permutations for which the log-rank statistic is (equally or) more
extreme than the log-rank stitistic for the data.
}
\description{
Get the p-value for the Heinze log-rank test.
Permutations are performed until the 95\% confidence interval for log-rank's p-value does not contain
0.05, and the size of the confidence interval is within 10\% of the estimated p-value.
}
