% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{get.logrank.pvalue}
\alias{get.logrank.pvalue}
\title{Get p-value from survdiff}
\usage{
get.logrank.pvalue(survdiff.res)
}
\arguments{
\item{survdiff.res}{return value from the function survidff from the library survival.}
}
\value{
survdiff's log-rank p-value.
}
\description{
Get log-rank's p-value from an object returned by survdiff.
}
