% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinze.R
\name{logrank.heinze}
\alias{logrank.heinze}
\title{Heinze log-rank test}
\usage{
logrank.heinze(
  formula = NULL,
  data,
  max.num.perms = NULL,
  num.perms.between.continue.tests = 1000,
  significance.threshold = NULL,
  confidence.relative.size = NULL,
  should.continue.func = NULL,
  mc.cores = 1,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A formula of the form time + event.type ~ group. time is the time of the event,
event.type is 1 if the event was censored and 0 otherwise (TODO), and group is the group of each sample.}

\item{data}{A data.frame in which the formula is evaluated. If formula is NULL, the first column of data
is used as time, the second as event.type, and the third as group.}

\item{max.num.perms}{The maximal number of permutations to perform.}

\item{num.perms.between.continue.tests}{The number of permutations performed between tests of whether
more permutations should be performed. Defaults to 1000.}

\item{significance.threshold}{If this number is set, permutations continue as long as the confidence interval
for the p-value contain this number (or until max.num.perms permutations are performed).}

\item{confidence.relative.size}{If this number is set, permutations continue as long as the 95\%
confidence interval for the p-value is "relatively small". That is, its boundaries are within a
confidence.relative.size factor of the estimated p-value. For example, if the estimated p-value is 0.01
and confidence.relative.size=2, then the 95\% confidence interval must be contained in [0.005, 0.02].}

\item{should.continue.func}{A function used to decide whether more permutations should be performed.
If this is set, confidence.relative.size and should.continue.func cannot be set. The function receives three
arguments: the number of permutations performed so far, the number of permutations with more extreme
log-rank statistic than the original, and the original p-value.
The function should return a binary value of whether more permutations should be performed.}

\item{mc.cores}{Number of cores to use to parallelize permutations.}

\item{seed}{Initialization of R's random seed, or NULL to not set the seed.}

\item{verbose}{True <=> info on the number of permutations should be printed.}
}
\value{
A list with the following entries:
pvalue - the estimation for log-rank's p-value.
conf.int - 95\% confidence interval for log-rank's p-value.
total.num.perms - number of permutations performed to estimate the p-value.
total.num.extreme.pvals - the number of permutations for which the log-rank statistic is (equally or) more
extreme than the log-rank stitistic for the data.
}
\description{
Get the p-value for the Heinze log-rank test.
}
\examples{
logrank.heinze(data=surv.data, max.num.perms=1000)
logrank.heinze(Survival + Death ~ cluster, data=surv.data, max.num.perms=1000)
}
